+++ main.go
@@ -5,4 +5,3 @@ import "github.com/qownnotes/qc/cmd"
 func main() {
 	cmd.Execute()
 }
-