+++ shell.nix
@@ -1,9 +1,10 @@
-{ pkgs ? import <nixpkgs> {} }:
-  pkgs.mkShell {
-    # nativeBuildInputs is usually what you want -- tools you need to run
-    nativeBuildInputs = with pkgs; [
-      just
-      go
-    ];
+{
+  pkgs ? import <nixpkgs> { },
+}:
+pkgs.mkShell {
+  # nativeBuildInputs is usually what you want -- tools you need to run
+  nativeBuildInputs = with pkgs; [
+    just
+    go
+  ];
 }
-